ai-service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0

# ML/AI libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
openai-whisper>=20231117
transformers>=4.35.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Battery simulation (Digital Twin)
pybamm>=23.9
casadi>=3.6.3

# Image/Video processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Database and caching
redis>=5.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=24.1.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
